commits from dev/hevce:
ROI enabled for gacc; BRC: inf loop protection at underflow
Gacc: 8K support (up to 8192x4320); bug fix in Refine32x32 kernel
refactored hme, cyclic buffer for meControls and postProcParams
fix failing tests
copy sei messages
GetFlBits
MvCost1RefLog
fix linux build (_BitScanReverse)

+++ _studio/mfx_lib/genx/h265_encode/include/genx_hevce_refine_me_p_common_old.h
@@ -49,7 +49,7 @@ void SumSad<32>(matrix_ref<ushort, 9, 16> partSad, vector_ref<uint, 9> sad)
 {
     matrix<ushort, 9, 8> t1 = SELECT_N_COLS(partSad, 0, 8) + SELECT_N_COLS(partSad, 8, 8);
     matrix<ushort, 9, 4> t2 = SELECT_N_COLS(t1, 0, 4) + SELECT_N_COLS(t1, 4, 4);
-    matrix<ushort, 9, 2> t3 = SELECT_N_COLS(t2, 0, 2) + SELECT_N_COLS(t2, 2, 2);
+    matrix<uint, 9, 2> t3 = SELECT_N_COLS(t2, 0, 2) + SELECT_N_COLS(t2, 2, 2);
     sad = SELECT_N_COLS(t3, 0, 1) + SELECT_N_COLS(t3, 1, 1);
 }
